chore: split channels into separate entities
